# doc/api/reference-v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "data-platform / OA 参照接口",
    "version": "1.0.0",
    "description": "开发契约，尚未部署。完全独立于DEE。全部样例为模拟数据，逻辑类型须在真实SQL发布时验证。"
  },
  "servers": [
    {
      "url": "https://platform.example.invalid/integration/openapi/v1",
      "description": "占位地址，联调时替换"
    }
  ],
  "security": [
    {
      "IntegrationKey": []
    }
  ],
  "paths": {
    "/status": {
      "get": {
        "operationId": "getIntegrationStatus",
        "summary": "检查协议入口和服务身份；不测试U8",
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "每次HTTP请求生成UUID，重试使用新值"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                },
                "example": {
                  "code": "OK",
                  "message": "成功",
                  "requestId": "83c87e30-8aba-4a82-b5d4-9233f52ead48",
                  "executionId": null,
                  "data": {
                    "status": "UP",
                    "apiVersion": "1.0.0",
                    "clientId": "oa-plugin-test"
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_ARGUMENT / FILTER_NOT_ALLOWED / SORT_NOT_ALLOWED / PARAMETER_NOT_ALLOWED",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "DATASOURCE_UNAVAILABLE / SERVICE_UNAVAILABLE",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reference/tasks": {
      "get": {
        "operationId": "listReferenceTasks",
        "summary": "列出已启用且有权限的参照任务",
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "每次HTTP请求生成UUID，重试使用新值"
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPageResponse"
                },
                "example": {
                  "code": "OK",
                  "message": "成功",
                  "requestId": "83c87e30-8aba-4a82-b5d4-9233f52ead48",
                  "executionId": null,
                  "data": {
                    "items": [
                      {
                        "taskCode": "U8_MATERIAL_REFERENCE",
                        "taskName": "物料参照",
                        "metadataVersion": "1"
                      }
                    ],
                    "total": 1,
                    "pageNum": 1,
                    "pageSize": 20,
                    "totalPages": 1
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_ARGUMENT / FILTER_NOT_ALLOWED / SORT_NOT_ALLOWED / PARAMETER_NOT_ALLOWED",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "DATASOURCE_UNAVAILABLE / SERVICE_UNAVAILABLE",
            "headers": {
              "X-Request-Id": {
                "description": "关联请求UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Cache-Control": {
                "description": "固定no-store",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "IntegrationKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Integration-Key",
        "description": "由平台发给OA插件服务端的专用凭据；不进入浏览器"
      }
    },
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "taskCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "taskName": {
            "type": "string",
            "minLength": 1
          },
          "metadataVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "taskCode",
          "taskName",
          "metadataVersion"
        ],
        "additionalProperties": false
      },
      "TaskPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "pageNum": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "pageNum",
          "pageSize",
          "total",
          "totalPages"
        ],
        "additionalProperties": false
      },
      "Status": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "UP"
            ]
          },
          "apiVersion": {
            "type": "string",
            "enum": [
              "1.0.0"
            ]
          },
          "clientId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "status",
          "apiVersion",
          "clientId"
        ],
        "additionalProperties": false
      },
...
      "StatusResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "OK"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "executionId": {
            "type": "string",
            "nullable": true,
            "enum": [
              null
            ]
          },
          "data": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "executionId",
          "data"
        ],
        "additionalProperties": false
      },
      "TaskPageResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "OK"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "executionId": {
            "type": "string",
            "nullable": true,
            "enum": [
              null
            ]
          },
          "data": {
            "$ref": "#/components/schemas/TaskPage"
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "executionId",
          "data"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_ARGUMENT",
              "FILTER_NOT_ALLOWED",
              "SORT_NOT_ALLOWED",
              "PARAMETER_NOT_ALLOWED",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "TASK_NOT_FOUND",
              "RESULT_SET_NOT_FOUND",
              "TASK_DISABLED",
              "METADATA_VERSION_MISMATCH",
              "RESULT_SCHEMA_MISMATCH",
              "REQUEST_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "DATASOURCE_UNAVAILABLE",
              "SERVICE_UNAVAILABLE",
              "QUERY_TIMEOUT",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "executionId": {
            "type": "string",
            "nullable": true,
            "minLength": 1
          },
          "data": {
            "type": "object",
            "nullable": true,
            "enum": [
              null
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "field",
                    "reason"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "retryable",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "executionId",
          "data",
          "error"
        ],
        "additionalProperties": false
      }
    }
  }
}